+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support for a new TTS service, `MiniMaxHttpTTSService`.
+  (see https://www.minimax.io/audio)
+
 - A new function `FrameProcessor.setup()` has been added to allow setting up
   frame processors before receiving a `StartFrame`. This is what's happening
   internally: `FrameProcessor.setup()` is called, `StartFrame` is pushed from